requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
boto3==1.39.4
botocore==1.39.4
certifi==2025.7.9
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.3
django-cors-headers==4.7.0
django-countries==7.6.1
django-environ==0.12.0
django-on-heroku==1.1.2
django-storages==1.14.4
django-taggit==6.1.0
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
ollama==0.5.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.9.0
